Fix: Restore event creation and simplify deletion logic
- Reverted problematic changes that blocked event creation for free users
- Simplified deleteEvent function to avoid complex state management
- Fixed sync logic for both Pro and Free users
- Removed unnecessary navigation health checks
- Updated support email to labs@brownster.com
- Fixed delete-all pattern in Firebase sync (now uses smart sync)
- Added proper error handling for deletion permissions
- Improved UI refresh with currentEvent reassignment

This commit restores stable functionality with:
✓ Event creation working for both Pro and Free users
✓ Proper sync between devices
✓ Smart Firebase sync without delete-all pattern
✓ Clear error messages for permission issues

+++ TheEqualizer/Models/Donation.swift
@@ -2,11 +2,11 @@ import Foundation
 
 struct Donation: Identifiable, Codable, Equatable {
     let id: UUID
-    var amount: Double
+    var amount: Decimal
     var notes: String
     let date: Date
     
-    init(id: UUID = UUID(), amount: Double, notes: String = "", date: Date = Date()) {
+    init(id: UUID = UUID(), amount: Decimal, notes: String = "", date: Date = Date()) {
         self.id = id
         self.amount = amount
         self.notes = notes
